remove warnings

+++ spirv_cross.cpp
@@ -3633,7 +3633,7 @@ void Compiler::make_constant_null(uint32_t id, uint32_t type)
 	}
 	else if (!constant_type.member_types.empty())
 	{
-		uint32_t member_ids = increase_bound_by(constant_type.member_types.size());
+		uint32_t member_ids = increase_bound_by((uint32_t)(constant_type.member_types.size()));
 		vector<uint32_t> elements(constant_type.member_types.size());
 		for (uint32_t i = 0; i < constant_type.member_types.size(); i++)
 		{
